requirements.txt
Flask==3.0.0
Flask-Session==0.5.0
mysql-connector-python==8.0.33
gunicorn==20.1.0